fix(repository): classify perception package inventory
